requirements.txt
numpy
pillow
opencv-python
requests
pydub
moviepy
scikit-image